Move api types into a single common file
