+++ common/api-types.ts
@@ -1,0 +1,9 @@
+export type ApiError = {
+  status: number,
+  message: string
+};
+
+export type UserToken = {
+  id: string,
+  email: string
+};
