--- common/users.ts
@@ -1,4 +1,0 @@
-export type UserToken = {
-  id: string,
-  email: string
-};
